divide modals into separate components
